refactored user ids and populated tasks

+++ controllers/taskController.js
@@ -6,9 +6,23 @@ const AppError = require('../utils/appError');
 
 exports.getAllTasks = factory.getAll(Task);
 exports.getTask = factory.getOne(Task);
-exports.createTask = factory.createOne(Task);
+// exports.createTask = factory.createOne(Task);
 // exports.updateTask = factory.updateOne(Task);
 // exports.deleteTask = factory.deleteOne(Task);
+
+exports.createTask = catchAsync(async (req, res, next) => {
+  const task = await Task.create({
+    ...req.body,
+    owner: req.user._id,
+  });
+
+  res.status(201).json({
+    status: 'success',
+    data: {
+      task,
+    },
+  });
+});
 
 exports.updateTask = catchAsync(async (req, res, next) => {
   const updates = Object.keys(req.body);
@@ -29,12 +43,17 @@ exports.updateTask = catchAsync(async (req, res, next) => {
     next(new AppError('Provide some data to update.', 400));
   }
 
-  const task = await Task.findByIdAndUpdate(req.params.id, req.body, {
+  let task = await Task.findById(req.params.id);
+
+  if (!task) return next(new AppError(`No task found with this id ${req.params.id}`, 404));
+
+  if (task.owner.toString() !== req.user.id)
+    return next(new AppError(`This task doesn't belong to current user`));
+
+  task = await Task.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
     runValidators: true,
   });
-
-  if (!task) return next(new AppError(`No task found with this id ${req.params.id}`, 404));
 
   res.status(200).json({
     status: 'success',
@@ -45,16 +64,18 @@ exports.updateTask = catchAsync(async (req, res, next) => {
 });
 
 exports.deleteTask = catchAsync(async (req, res, next) => {
-  const task = await Task.findByIdAndDelete(req.params.id);
-  if (!task) return next(new AppError('No document found with that ID', 404));
-  let count = await Task.countDocuments({ completed: true });
-  if (!count) count = 'No documents match this criteria';
+  let task = await Task.findById(req.params.id);
+  if (!task) return next(new AppError(`No task found with this id ${req.params.id}`, 404));
+
+  if (task.owner.toString() !== req.user.id)
+    return next(new AppError(`This task doesn't belong to current user`));
+
+  task = await Task.findByIdAndDelete(req.params.id);
 
   res.status(200).json({
     status: 'success',
     data: {
       task,
-      count,
     },
   });
 });
